refact: aplicando responsabilidad unica en flujo cancelar
