+++ src/templates/flujos/cancelarCita/index.ts
@@ -1,3 +1,6 @@
 export * from './step1CancelarCita';
-export * from './step3CancelarCita';
 export * from './stepOpcionReprogramar';
+export * from './step5CancelarCita';
+export * from './step6CancelarCita';
+export * from './step7CancelarCita';
+export * from './stepConfirmaCancelarCita';
